requirements/requirements-5.1.6.txt
AccessControl==3.0.14
Acquisition==4.4.2
BTrees==4.4.1
Chameleon==2.25
DateTime==4.2
DocumentTemplate==2.13.4
ExtensionClass==4.3.0
Markdown==2.6.9
Missing==3.2
MultiMapping==3.1
Persistence==2.13.2
Pillow==5.2.0
Plone==5.1.5
Products.ATContentTypes==2.3.9
Products.Archetypes==1.15.6
Products.BTreeFolder2==2.14.0
Products.CMFCore==2.2.12
Products.CMFDiffTool==3.2.2
Products.CMFDynamicViewFTI==5.0.0
Products.CMFEditions==3.2.2
Products.CMFFormController==3.1.5
Products.CMFPlacefulWorkflow==1.7.6
Products.CMFPlone==5.1.5
Products.CMFQuickInstallerTool==3.0.16
Products.CMFUid==2.2.2
Products.DCWorkflow==2.2.5
Products.DateRecurringIndex==2.1
Products.ExtendedPathIndex==3.4.1
Products.ExternalEditor==1.1.3
Products.ExternalMethod==2.13.1
Products.GenericSetup==1.8.10
Products.MIMETools==2.13.0
Products.MailHost==2.13.2
Products.MimetypesRegistry==2.1.5
Products.OFSP==2.13.2
Products.PlacelessTranslationService==2.0.7
Products.PlonePAS==5.1.0
Products.PluggableAuthService==1.11.0
Products.PluginRegistry==1.4
Products.PortalTransforms==3.1.5
Products.PythonScripts==2.13.2
Products.ResourceRegistries==3.0.7
Products.Sessions==3.0
Products.StandardCacheManagers==2.13.1
Products.TemporaryFolder==3.0
Products.ZCTextIndex==2.13.5
Products.ZCatalog==3.0.3
Products.ZSQLMethods==2.13.5
Products.ZopeVersionControl==1.1.4
Products.contentmigration==2.1.19
Products.statusmessages==5.0.4
Products.validation==2.1.1
PyJWT==1.7.1
Record==2.13.0
RestrictedPython==3.6.0
Unidecode==0.4.1
ZConfig==3.1.0
ZEO==5.1.1
ZODB==5.3.0
ZODB3==3.11.0
ZServer==3.0
Zope2==2.13.27
ZopeUndo==4.3
archetypes.multilingual==3.0.6
archetypes.schemaextender==2.1.8
borg.localrole==3.1.6
collective.monkeypatcher==1.2
collective.recipe.plonesite==1.9.5
cssselect==1.0.1
decorator==4.1.2
diazo==1.2.11
docutils==0.14
enum34==1.1.6
feedparser==5.2.1
five.customerize==1.1
five.globalrequest==1.0
five.intid==1.1.2
five.localsitemanager==2.0.6
five.pt==2.2.5
functools32==3.2.3.post2
future==0.16.0
futures==3.1.1
icalendar==4.0.3
initgroups==4.0
jsonschema==2.6.0
lxml==4.2.1
mailinglogger==3.8.0
mechanize==0.2.5
mockup==2.7.7
mr.developer==1.38
pathlib==1.0.1
persistent==4.2.4.2
piexif==1.0.13
plone.alterego==1.1.3
plone.api==1.9.1
plone.app.blob==1.7.4
plone.app.caching==1.2.23
plone.app.collection==1.2.7
plone.app.content==3.5.5
plone.app.contentlisting==1.3.3
plone.app.contentmenu==2.2.4
plone.app.contentrules==4.0.18
plone.app.contenttypes==1.4.15
plone.app.controlpanel==3.0.4
plone.app.customerize==1.3.9
plone.app.dexterity==2.5.2
plone.app.discussion==3.0.8
plone.app.event==3.2.1
plone.app.folder==1.2.6
plone.app.i18n==3.0.5
plone.app.imaging==2.0.7
plone.app.intid==1.1.3
plone.app.iterate==3.3.10
plone.app.layout==2.8.3
plone.app.linkintegrity==3.3.7
plone.app.locales==5.1.12
plone.app.lockingbehavior==1.0.6
plone.app.multilingual==5.2.3
plone.app.portlets==4.3.4
plone.app.querystring==1.4.11
plone.app.redirector==1.3.7
plone.app.registry==1.7.2
plone.app.relationfield==1.4.3
plone.app.textfield==1.2.11
plone.app.theming==2.0.5
plone.app.upgrade==2.0.20
plone.app.users==2.4.3
plone.app.uuid==1.2
plone.app.versioningbehavior==1.3.6
plone.app.viewletmanager==2.0.12
plone.app.vocabularies==4.0.7
plone.app.widgets==2.4.1
plone.app.workflow==3.0
plone.app.z3cform==3.0.8
plone.autoform==1.8.1
plone.batching==1.1.5
plone.behavior==1.2.1
plone.browserlayer==2.2.2
plone.cachepurging==1.0.16
plone.caching==1.2.1
plone.contentrules==2.0.9
plone.dexterity==2.6.2
plone.event==1.4.0
plone.folder==2.0.2
plone.formwidget.namedfile==2.0.7
plone.formwidget.recurrence==2.1.3
plone.i18n==3.0.7
plone.indexer==1.0.5
plone.intelligenttext==3.0.0
plone.keyring==3.1.0
plone.locking==2.2.3
plone.memoize==1.2.3
plone.namedfile==4.2.6
plone.outputfilters==3.0.5
plone.portlet.collection==3.3.1
plone.portlet.static==3.1.3
plone.portlets==2.3
plone.protect==3.1.4
plone.recipe.zope2instance==4.4.1
plone.registry==1.1.5
plone.resource==2.0.2
plone.resourceeditor==2.1.3
plone.rfc822==1.1.4
plone.scale==3.0.3
plone.schema==1.2.0
plone.schemaeditor==2.0.20
plone.session==3.7.1
plone.stringinterp==1.3.1
plone.subrequest==1.8.6
plone.supermodel==1.4.0
plone.synchronize==1.0.3
plone.theme==3.0.6
plone.transformchain==2.0.1
plone.uuid==1.0.5
plone.z3cform==0.9.1
plonetheme.barceloneta==2.0.0a1
ply==3.4
pyScss==1.3.5
python-dateutil==2.7.3
python-gettext==3.0
pytz==2017.3
repoze.xmliter==0.6
roman==1.4.0
simplejson==3.12.0
six==1.10.0
slimit==0.8.1
sourcecodegen==0.6.14
tempstorage==4.0.1
transaction==2.1.2
trollius==2.1
z3c.caching==2.0a1
z3c.form==3.6
z3c.formwidget.query==0.17
z3c.objpath==1.1
z3c.pt==3.0.0a1
z3c.relationfield==0.7
z3c.zcmlhook==1.0b1
zExceptions==2.13.0
zLOG==3.0
zc.lockfile==1.2.1
zc.recipe.egg==2.0.5
zc.relation==1.0
zdaemon==4.2.0
zodbpickle==0.7.0
zope.annotation==3.5.0
zope.app.locales==3.7.5
zope.app.publication==3.12.0
zope.authentication==3.7.1
zope.broken==3.6.0
zope.browser==2.1.0
zope.browsermenu==4.2
zope.browserpage==4.1.0
zope.browserresource==4.1.0
zope.cachedescriptors==3.5.1
zope.component==4.4.1
zope.componentvocabulary==1.0.1
zope.configuration==3.7.4
zope.container==3.11.2
zope.contentprovider==3.7.2
zope.contenttype==4.2.0
zope.copy==3.5.0
zope.datetime==3.4.1
zope.deferredimport==3.5.3
zope.deprecation==4.3.0
zope.dottedname==4.2
zope.error==3.7.4
zope.event==3.5.2
zope.exceptions==3.6.2
zope.filerepresentation==3.6.1
zope.globalrequest==1.2
zope.i18n==4.2.0
zope.i18nmessageid==4.1.0
zope.interface==4.4.3
zope.intid==3.7.2
zope.keyreference==3.6.4
zope.lifecycleevent==3.6.2
zope.location==3.9.1
zope.pagetemplate==4.2.1
zope.processlifetime==1.0
zope.proxy==4.3.0
zope.ptresource==3.9.0
zope.publisher==4.3.2
zope.ramcache==2.2.0
zope.schema==4.5.0
zope.security==4.1.1
zope.sendmail==3.7.5
zope.sequencesort==3.4.0
zope.site==3.9.2
zope.size==3.4.1
zope.structuredtext==3.5.1
zope.tal==3.5.2
zope.tales==3.5.3
zope.testbrowser==3.11.1
zope.testing==3.9.7
zope.traversing==4.1.0
zope.untrustedpython==4.0.0
zope.viewlet==3.7.2
